Add 15-minute and 30-minute cache timeout options for API providers
Added 0.25h (15min) and 0.5h (30min) cache timeout options to all 4
provider dropdowns. Fixed parseInt→parseFloat in updateProviderSettings
so fractional hour values are preserved correctly.

Shorter cache windows pair well with the new 1-day intraday sparkline
(STACK-66), giving users more frequent data points for richer charts.

+++ js/api.js
@@ -360,7 +360,7 @@ const updateProviderSettings = (provider) => {
   const cacheSelect = document.getElementById(`cacheTimeout_${provider}`);
   if (cacheSelect) {
     if (!config.cacheTimeouts) config.cacheTimeouts = {};
-    config.cacheTimeouts[provider] = parseInt(cacheSelect.value);
+    config.cacheTimeouts[provider] = parseFloat(cacheSelect.value);
   }
 
   saveApiConfig(config);
